pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diskalert"
version = "0.0.1"
description = "Disk Alert application"
readme = "README.rst"
authors = [{name = "Heechul Kim", email = "jijisa@iorchard.net"}]
dependencies = [
    "python-dotenv"
]
[project.scripts]
diskalert = "diskalert.__main__:main"
